@@ -11,14 +11,15 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~7.2.3",
-    "@angular/common": "~7.2.3",
-    "@angular/compiler": "~7.2.3",
-    "@angular/core": "~7.2.3",
-    "@angular/forms": "~7.2.3",
-    "@angular/platform-browser": "~7.2.3",
-    "@angular/platform-browser-dynamic": "~7.2.3",
-    "@angular/router": "~7.2.3",
+    "@angular/animations": "~8.1.2",
+    "@angular/common": "~8.1.2",
+    "@angular/compiler": "~8.1.2",
+    "@angular/core": "~8.1.2",
+    "@angular/elements": "^8.1.2",
+    "@angular/forms": "~8.1.2",
+    "@angular/platform-browser": "~8.1.2",
+    "@angular/platform-browser-dynamic": "~8.1.2",
+    "@angular/router": "~8.1.2",
     "@ng-bootstrap/ng-bootstrap": "^4.0.2",
     "body-parser": "^1.18.3",
     "bootstrap": "^4.2.1",
@@ -27,19 +28,20 @@
     "express": "^4.16.4",
     "moment": "^2.24.0",
     "node-mailjet": "^3.3.1",
-    "rxjs": "~6.4.0",
+    "rxjs": "~6.5.2",
     "tslib": "^1.9.0",
-    "zone.js": "~0.8.26"
+    "zone.js": "~0.9.1",
+    "document-register-element": "^1.7.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.13.0",
-    "@angular/cli": "~7.3.0",
-    "@angular/compiler-cli": "~7.2.3",
-    "@angular/language-service": "~7.2.3",
+    "@angular-devkit/build-angular": "~0.801.2",
+    "@angular/cli": "~8.1.2",
+    "@angular/compiler-cli": "~8.1.2",
+    "@angular/language-service": "~8.1.2",
     "@types/jasmine": "~3.3.8",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~10.12.21",
-    "codelyzer": "~4.5.0",
+    "codelyzer": "^5.0.1",
     "jasmine-core": "~3.3.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~4.0.0",
@@ -50,6 +52,6 @@
     "protractor": "~5.4.0",
     "ts-node": "~8.0.2",
     "tslint": "~5.12.1",
-    "typescript": "~3.2.4"
+    "typescript": "~3.4.5"
   }
 }
